# Conflicts:
#	db/schema.rb
#	db/seeds.rb

+++ db/seeds.rb
@@ -1,7 +1,9 @@
 require 'csv' 
 
-Member.destroy_all
-Product.destroy_all
+if Rails.env.development?
+  Member.destroy_all
+  Product.destroy_all
+end
 
 CSV.foreach(Rails.root.join('db/products.csv'), headers: true, header_converters: :symbol) do |row|
   product = Product.find_or_create_by(name: row[:name]) do |product|
@@ -15,7 +17,8 @@ CSV.foreach(Rails.root.join('db/products.csv'), headers: true, header_converters
     product.solution = row[:solution]
     product.cover_picture = row[:cover_picture] ? "https://res.cloudinary.com/wagon/" + row[:cover_picture] : nil
     product.customer_segment = row[:customer_segment]
-    product.batch = "#{row[:city]} #row#{[:slug]}"
+    product.batch = "#{row[:city]} ##{row[:slug]}"
+    product.batch_number = row[:slug]
   end
   product.members.create(first_name: row[:first_name], last_name: row[:last_name])
 end
